Should now be able to boot qd86.bin. Fixed compiler to 32-bit and got rid of unsigned long long tripping up x86 gcc.

+++ src/alpha.h
@@ -10,7 +10,7 @@
 #define INT64_DIGITS_MAX 11
 #define INT_DIGITS_MAX 10
 
-typedef int64_t UTOA_T; /*!< Biggest type utoa can convert to string. */
+typedef uint32_t UTOA_T; /*!< Biggest type utoa can convert to string. */
 #define UTOA_DIGITS_MAX INT64_DIGITS_MAX
 
 #include "mem.h" /* This uses UTOA_DIGITS_MAX for mvalue. */
